NEON-1009 Fix Hive's lowercasing of field names
The Hive engine (which Neon uses to connect to Spark SQL) is case insensitive,
and all field/column names are converted to lower case. This means that when
there is an "AS" in the select statement (e.g., "count(foo) AS FooCount"), then
Hive will return the result in lower case (e.g., "foocount").  This code detects
that case and fixes the response to have the capitalization that the user
requested.