NEON-1368 Fix HTTP codes for missing databases/tables
Try to return 502 (Bad Gateway) when a connection to a database fails because
the requested server was not there. This is indicated by
NoNodeAvailableException in ElasticSearch, or when an UnknownHostException or
ConnectionException causes an otherwise unrecognized exception.

Mongo tends to just let bad hostnames timeout and throws a
MongoTimeoutException, so return 504 (Gateway Timeout).

If we know that the requested database or collection doesn't exist, then return
404 (Not Found).  In Mongo, it's important to check databases because if you ask
for one that doesn't exist, Mongo will automatically create it.

The Spark JDBC connector just throws SQLExceptions for databases and tables that
don't exist, so just return 500 (Internal Server Error) since it's hard to
determine the exact cause without parsing exception messages.